+++ example/with-paths/app.ts
@@ -1,0 +1,3 @@
+import inc from './math';
+
+console.log('two', inc(1));