+++ example/with-paths/utils/src/to-ignore.ts
@@ -1,0 +1,2 @@
+// Should be ignored due to --ignorePaths option
+export const unused_fun = () => 0;