+++ ispec/run.sh
@@ -7,3 +7,7 @@ popd
 pushd ../example/tsx
 npm i > /dev/null && ../../ispec/_run-and-check-exit-code.sh
 popd
+
+pushd ../example/with-paths
+npm i > /dev/null && ../../ispec/_run-and-check-exit-code.sh
+popd