+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 
 ### Changed
+- Fix the --ignorePaths option (it was incorrectly filtering the parsed files, instead of filtering the output)
 
 ## [2.1.0] - 20 Oct 2019
 ### Added